Vídeo 2: Autorización
Segundo vídeo de la lista completado, el código es el mismo, pero por problema de mongo hay que probar las operaciones con Postman

+++ backend/controllers/goalController.js
@@ -1,12 +1,14 @@
 'use strict'
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
+ 
 
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -22,6 +24,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id, 
   });
 
   res.status(200).json(goal);
@@ -37,6 +40,20 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error('Goal not found');
+  }
+
+  const user = await User.findById(req.user.id)
+
+  // Comprobar si hay usuario
+  if(!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // Asegurarse de que el usuario autenticado coincide con su goal respectivo
+  if(goal.user.toString !== user.id){
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   const updateGoal = await Goal.findByIdAndUpdate(id, req.body, {
@@ -56,6 +73,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error('Goal not found');
+  }
+
+  const user = await User.findById(req.user.id)
+
+  // Comprobar si hay usuario
+  if(!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // Asegurarse de que el usuario autenticado coincide con su goal respectivo
+  if(goal.user.toString !== user.id){
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await Goal.findByIdAndDelete(id);
